break down server module
